pyproject.toml
[project]
name = "er-helper"
version = "0.1.0"
description = "Generate Mermaid ER diagrams from a MySQL database schema"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy",
    "pymysql",
    "python-dotenv",
    "pyperclip",
]
